Upgraded to wicket 6
git-svn-id: https://superfly.googlecode.com/svn/branches/R1.5@929 11a29e22-99e0-9bb7-d19e-9a0993c63ea3

+++ superfly-integration-test/src/test/java/com/payneteasy/superfly/dao/ActionDaoTest.java
@@ -60,7 +60,7 @@ public class ActionDaoTest extends AbstractDaoTest {
 
     @Test
 	public void testGetActionCount() {
-		int count = actionDao.getActionCount(null, null, null);
+		long count = actionDao.getActionCount(null, null, null);
         Assert.assertTrue("Must get some action", count > 0);
 		actionDao.getActionCount("someActionName", "someActionDescription",
 				"1,2");
